+++ tests/integration/authy/v1/service/entity/factor/test_challenge.py
@@ -49,7 +49,6 @@ class ChallengeTestCase(IntegrationTestCase):
                 "responded_reason": "none",
                 "details": "Hi! Mr. John Doe, would you like to sign up?",
                 "hidden_details": "Hidden details about the sign up",
-                "factor_strength": "low",
                 "factor_type": "sms",
                 "url": "https://authy.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Factors/YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
@@ -127,7 +126,6 @@ class ChallengeTestCase(IntegrationTestCase):
                 "responded_reason": "none",
                 "details": "details",
                 "hidden_details": "hidden_details",
-                "factor_strength": "low",
                 "factor_type": "sms",
                 "url": "https://authy.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Factors/YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
@@ -160,7 +158,6 @@ class ChallengeTestCase(IntegrationTestCase):
                 "responded_reason": "none",
                 "details": "details",
                 "hidden_details": "hidden_details",
-                "factor_strength": "low",
                 "factor_type": "sms",
                 "url": "https://authy.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Factors/YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
@@ -209,7 +206,6 @@ class ChallengeTestCase(IntegrationTestCase):
                 "responded_reason": "none",
                 "details": "Hi! Mr. John Doe, would you like to sign up?",
                 "hidden_details": "Hidden details about the sign up",
-                "factor_strength": "low",
                 "factor_type": "sms",
                 "url": "https://authy.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Factors/YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
@@ -242,7 +238,6 @@ class ChallengeTestCase(IntegrationTestCase):
                 "responded_reason": "none",
                 "details": "Hi! Mr. John Doe, would you like to sign up?",
                 "hidden_details": "Hidden details about the sign up",
-                "factor_strength": "low",
                 "factor_type": "sms",
                 "url": "https://authy.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Factors/YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }