+++ tests/integration/authy/v1/service/entity/test_factor.py
@@ -50,7 +50,6 @@ class FactorTestCase(IntegrationTestCase):
                 "date_updated": "2015-07-30T20:00:00Z",
                 "friendly_name": "friendly_name",
                 "status": "unverified",
-                "factor_strength": "low",
                 "factor_type": "sms",
                 "url": "https://authy.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Factors/YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -122,7 +121,6 @@ class FactorTestCase(IntegrationTestCase):
                 "date_updated": "2015-07-30T20:00:00Z",
                 "friendly_name": "friendly_name",
                 "status": "unverified",
-                "factor_strength": "low",
                 "factor_type": "sms",
                 "url": "https://authy.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Factors/YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {
@@ -194,7 +192,6 @@ class FactorTestCase(IntegrationTestCase):
                         "date_updated": "2015-07-30T20:00:00Z",
                         "friendly_name": "friendly_name",
                         "status": "unverified",
-                        "factor_strength": "low",
                         "factor_type": "sms",
                         "url": "https://authy.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Factors/YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "links": {
@@ -250,7 +247,6 @@ class FactorTestCase(IntegrationTestCase):
                 "date_updated": "2015-07-30T20:00:00Z",
                 "friendly_name": "friendly_name",
                 "status": "verified",
-                "factor_strength": "low",
                 "factor_type": "sms",
                 "url": "https://authy.twilio.com/v1/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Factors/YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "links": {