+++ tests/integration/preview/trusted_comms/test_current_call.py
@@ -18,10 +18,9 @@ class CurrentCallTestCase(IntegrationTestCase):
         self.holodeck.mock(Response(500, ''))
 
         with self.assertRaises(TwilioException):
-            self.client.preview.trusted_comms.current_calls().fetch(twilio_sandbox_mode="twilio_sandbox_mode", x_xcnam_sensitive_phone_number_from="x_xcnam_sensitive_phone_number_from", x_xcnam_sensitive_phone_number_to="x_xcnam_sensitive_phone_number_to")
+            self.client.preview.trusted_comms.current_calls().fetch(x_xcnam_sensitive_phone_number_from="x_xcnam_sensitive_phone_number_from", x_xcnam_sensitive_phone_number_to="x_xcnam_sensitive_phone_number_to")
 
         headers = {
-            'Twilio-Sandbox-Mode': "twilio_sandbox_mode",
             'X-Xcnam-Sensitive-Phone-Number-From': "x_xcnam_sensitive_phone_number_from",
             'X-Xcnam-Sensitive-Phone-Number-To': "x_xcnam_sensitive_phone_number_to",
         }