+++ twilio/rest/preview/trusted_comms/cps.py
@@ -121,21 +121,16 @@ class CpsContext(InstanceContext):
         self._solution = {}
         self._uri = '/CPS'.format(**self._solution)
 
-    def fetch(self, twilio_sandbox_mode=values.unset,
-              x_xcnam_sensitive_phone_number=values.unset):
+    def fetch(self, x_xcnam_sensitive_phone_number=values.unset):
         """
         Fetch the CpsInstance
 
-        :param unicode twilio_sandbox_mode: Optional header to mock all voice dependencies.
         :param unicode x_xcnam_sensitive_phone_number: Phone number to retrieve CPS.
 
         :returns: The fetched CpsInstance
         :rtype: twilio.rest.preview.trusted_comms.cps.CpsInstance
         """
-        headers = values.of({
-            'Twilio-Sandbox-Mode': twilio_sandbox_mode,
-            'X-Xcnam-Sensitive-Phone-Number': x_xcnam_sensitive_phone_number,
-        })
+        headers = values.of({'X-Xcnam-Sensitive-Phone-Number': x_xcnam_sensitive_phone_number, })
 
         payload = self._version.fetch(method='GET', uri=self._uri, headers=headers, )
 
@@ -214,21 +209,16 @@ class CpsInstance(InstanceResource):
         """
         return self._properties['url']
 
-    def fetch(self, twilio_sandbox_mode=values.unset,
-              x_xcnam_sensitive_phone_number=values.unset):
+    def fetch(self, x_xcnam_sensitive_phone_number=values.unset):
         """
         Fetch the CpsInstance
 
-        :param unicode twilio_sandbox_mode: Optional header to mock all voice dependencies.
         :param unicode x_xcnam_sensitive_phone_number: Phone number to retrieve CPS.
 
         :returns: The fetched CpsInstance
         :rtype: twilio.rest.preview.trusted_comms.cps.CpsInstance
         """
-        return self._proxy.fetch(
-            twilio_sandbox_mode=twilio_sandbox_mode,
-            x_xcnam_sensitive_phone_number=x_xcnam_sensitive_phone_number,
-        )
+        return self._proxy.fetch(x_xcnam_sensitive_phone_number=x_xcnam_sensitive_phone_number, )
 
     def __repr__(self):
         """